Support more file format for flag files
