+++ go.mod
@@ -8,6 +8,7 @@ require (
 	github.com/blang/semver v3.5.1+incompatible // indirect
 	github.com/google/go-cmp v0.5.4
 	github.com/nikunjy/rules v0.0.0-20200120082459-0b7c4dc9dc86
+	github.com/pelletier/go-toml v1.8.1
 	github.com/stretchr/testify v1.7.0
 	gopkg.in/yaml.v2 v2.4.0 // indirect
 	gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c
