+++ testdata/flag-config.json
@@ -1,0 +1,16 @@
+{
+  "test-flag": {
+    "rule": "key eq \"random-key\"",
+    "percentage": 100,
+    "true": true,
+    "false": false,
+    "default": false
+  },
+  "test-flag2": {
+    "rule": "key eq \"not-a-key\"",
+    "percentage": 100,
+    "true": true,
+    "false": false,
+    "default": false
+  }
+}
